requirements.txt
asgiref==3.6.0
coverage==7.2.3
Django==4.2
django-js-asset==2.0.0
djangorestframework==3.14.0
execnet==1.9.0
factory-boy==3.2.1
Faker==12.0.1
iniconfig==2.0.0
jedi==0.18.2
mixer==7.2.2
packaging==23.1
parso==0.8.3
Pillow==9.5.0
pluggy==1.0.0
pudb==2022.1.3
Pygments==2.15.1
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-xdist==3.2.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
toposort==1.10
urwid==2.1.2
urwid-readline==0.13
